Complete VALIDATED state (run-000004)
- All 6 acceptance gates PASS
- G1 Source Identity: commit 857610125 verified
- G2 Binary Integrity: SHA256 hashes match
- G3 Execution Completion: exit code 0
- G4 Numerical Correctness: max_error=0.0
- G5 Determinism: repeat execution matches (seed=42)
- G6 Evidence Closure: all artifacts present

Validation Result: PASS (100% gate completion)
Evidence chain: source -> binary -> execution -> output hash

Lifecycle: EXECUTED -> VALIDATED complete
Next: VERIFIED state with evidence seal

+++ src/integration/val016_source_repair_v2.cpp
@@ -113,7 +113,7 @@ struct RepairPatch {
     size_t offset;
     std::string before;
     std::string after;
-    float confidence;
+    double confidence;
     std::string reason;
     
     nlohmann::json toJson() const {
@@ -186,7 +186,7 @@ enum class RepairType {
 struct RepairClassification {
     RepairType type;
     std::string description;
-    float confidence;
+    double confidence;
     std::string reason;
 };
